+++ index.tsx
@@ -1,11 +1,11 @@
 import React, { useRef, useEffect, CSSProperties } from 'react';
 import Quill, { RangeStatic, Sources } from 'quill';
-// import QuillBetterTable from 'quill-better-table';
 import Delta from 'quill-delta';
 import IconUndo from 'quill/assets/icons/undo.svg';
 import IconRedo from 'quill/assets/icons/redo.svg';
 import {
   highlightInit,
+  Image,
   ImageDrop,
   ImageResize,
   MagicUrl,
@@ -14,11 +14,11 @@ import {
   MarkdownShortcuts,
 } from './modules/index';
 import { imageUpload, linkHandler, undoHandler, redoHandler } from './modules/toolbarHandler';
-import { optionDisableToggle, setContent } from './utils';
+import { isUrl, optionDisableToggle, saveLink, setContent } from './utils';
 import 'quill/dist/quill.snow.css';
-// import 'quill-better-table/dist/quill-better-table.css';
 import './assets/richTextEditor.less';
 import './assets/modules.less';
+import normalizeUrl from 'normalize-url';
 
 interface IBetterTable {
   operationMenu?: {
@@ -77,6 +77,7 @@ type Title = {
   placeholder?: string;
   onChange?: (title: string) => void;
   defaultValue?: string;
+  value?: string;
   onFocus?: (e: React.FocusEvent<HTMLInputElement>) => void;
   onBlur?: (e: React.FocusEvent<HTMLInputElement>) => void;
 };
@@ -212,7 +213,15 @@ const RichTextEditor = (props: IEditorProps) => {
 
       // 默认添加图片缩放功能
       if (imageResize) {
-        quillModules.current.imageResize = typeof imageResize !== 'boolean' ? imageResize : {};
+        quillModules.current.imageResize =
+          imageResize === false
+            ? imageResize
+            : {
+                onUpdate: () => {
+                  console.log('img update');
+                },
+                ...(typeof imageResize === 'object' ? imageResize : null),
+              };
       }
       // 默认图片拖拽/复制到富文本
       if (imageDrop) {
@@ -292,6 +301,7 @@ const RichTextEditor = (props: IEditorProps) => {
       );
     }
 
+    Quill.register(Image, true);
     Quill.register(
       {
         'modules/imageResize': ImageResize,
@@ -367,6 +377,40 @@ const RichTextEditor = (props: IEditorProps) => {
                 'table-cell-line': false, // 在table中不触发有序列表
               },
             },
+            // bugfix: 当最开始是code块、list、引用块时，无法使用Backspace删除样式
+            'code backspace': {
+              key: 'Backspace',
+              format: ['code-block', 'list', 'blockquote'],
+              handler(
+                range: RangeStatic,
+                context: {
+                  line: { parent: { cachedText?: string } };
+                  suffix: string;
+                  prefix: string;
+                  offset: number;
+                },
+              ) {
+                const quill = quillRef.current;
+                if (quill) {
+                  const [line] = quill.getLine(range.index);
+                  const isEmpty = !line.children.head.text || line.children.head.text.trim() === '';
+                  const format = quill.getFormat(range);
+                  const allCode = context?.line?.parent?.cachedText;
+                  // 当是起始，代码块且整块中已无字符，或引用/列表且当前行为空，去除当前行格式；其他情况执行默认Backspace的handler
+                  if (
+                    range.index === 0 &&
+                    context.suffix === '' &&
+                    ((format['code-block'] && (allCode === '\n' || allCode === '')) ||
+                      (!format['code-block'] && isEmpty))
+                  ) {
+                    quill.removeFormat(range.index, range.length);
+                    return false;
+                  }
+                }
+
+                return true;
+              },
+            },
           },
         },
 
@@ -418,14 +462,7 @@ const RichTextEditor = (props: IEditorProps) => {
               (document.querySelector('a.ql-action') as HTMLAnchorElement).onclick = () => {
                 if (quillRef.current) {
                   quillRef.current.deleteText(range.index - offset, leaf.length());
-                  quillRef.current.insertText(
-                    range.index - offset,
-                    (document.getElementById('link-words') as HTMLInputElement).value,
-                    'link',
-                    (document.getElementById('link-url') as HTMLInputElement).value,
-                    'user',
-                  );
-                  if (quillRef.current.theme) quillRef.current.theme.tooltip.hide();
+                  saveLink(quillRef.current, range.index - offset);
                 }
               };
             }
@@ -514,6 +551,7 @@ const RichTextEditor = (props: IEditorProps) => {
             config.onChange && config.onChange(e.target.value)
           }
           defaultValue={config.defaultValue}
+          value={config.value}
           onFocus={config.onFocus}
           onBlur={config.onBlur}
         />