+++ lib/modules/index.d.ts
@@ -1,0 +1,9 @@
+import { ImageDrop } from './imagePasteDrop';
+import ImageResize from './imageResize';
+import { MagicUrl } from './magic-url';
+import MarkdownShortcuts from './markdown-shortcuts';
+import ToolbarTable from './toolbar-table';
+import highlightInit from './highlight';
+import QuillBetterTable from './quill-better-table/quill-better-table';
+import Image from './image';
+export { highlightInit, ImageDrop, Image, ImageResize, MagicUrl, MarkdownShortcuts, ToolbarTable, QuillBetterTable, };