+++ modules/toolbarHandler.js
@@ -1,19 +1,20 @@
 import Quill from 'quill';
-import { isUrl, isEmail } from '../utils';
+import normalizeUrl from 'normalize-url';
+import { isUrl, isEmail, saveLink } from '../utils';
 import close from '../assets/icons/icon_close.svg';
 
 export function imageUpload(quill, imgUploadApi, uploadSuccCB, uploadFailCB) {
   quill.disable();
 
   let modal = quill.container.querySelector('div.image-tooltip');
-  let fileInput;
+  let fileInput = null;
 
   if (modal === null) {
     modal = document.createElement('div');
     modal.classList.add('image-tooltip');
     modal.innerHTML = `
       <input type="file" class="ql-image" accept="image/png, image/gif, image/jpeg, image/bmp, image/x-icon" />
-      <div class="image-upload-btn">上传图片</div>
+      <div class="image-upload-btn">选择本地图片</div>
       <p class="image-url-label">您也可以输入网络图片URL</p>
       <div class="image-url-form"><input type="text" placeholder="http://example.com/image.png" class="image-url-input" /><span class="url-submit">插入</span></div>
       <div class="image-close">${close}</div>
@@ -26,12 +27,12 @@ export function imageUpload(quill, imgUploadApi, uploadSuccCB, uploadFailCB) {
       quill.enable(true);
       e.stopPropagation();
     });
-    modal.querySelector('.url-submit').addEventListener('click', (e) => {
+    modal.querySelector('.url-submit').onclick = (e) => {
       e.stopPropagation();
       const url = urlInput.value;
       if (url && isUrl(url)) {
         const range = quill.getSelection(true);
-        quill.editor.insertEmbed(range.index, 'image', url);
+        quill.editor.insertEmbed(range.index, 'image', normalizeUrl(url));
         quill.setSelection(range.index + 1, Quill.sources.SILENT);
         modal.classList.add('ql-hidden');
         urlInput.value = '';
@@ -39,17 +40,19 @@ export function imageUpload(quill, imgUploadApi, uploadSuccCB, uploadFailCB) {
       } else {
         console.log('请输入正确URL');
       }
-    });
-    modal.querySelector('.image-upload-btn').addEventListener('click', (e) => {
+    };
+    modal.querySelector('.image-upload-btn').onclick = (e) => {
       e.stopPropagation();
       fileInput.click();
-    });
+    };
   } else {
     modal.classList.remove('ql-hidden');
     fileInput = modal.querySelector('input.ql-image[type=file]');
   }
 
-  fileInput.addEventListener('change', () => {
+  // bugfix：多次插入本地图片时，由于每次都是addEventListener添加事件导致多次插入
+  // fileInput.addEventListener('change', () => {
+  fileInput.onchange = () => {
     const { files } = fileInput;
     const range = quill.getSelection(true);
 
@@ -75,7 +78,7 @@ export function imageUpload(quill, imgUploadApi, uploadSuccCB, uploadFailCB) {
         uploadFailCB(error);
       },
     );
-  });
+  };
 }
 
 export function linkHandler(value) {
@@ -93,14 +96,7 @@ export function linkHandler(value) {
       document.getElementById('link-words').value = '';
       document.getElementById('link-url').value = '';
       document.querySelector('a.ql-action').onclick = () => {
-        this.quill.insertText(
-          selection.index,
-          document.getElementById('link-words').value,
-          'link',
-          document.getElementById('link-url').value,
-          'user',
-        );
-        this.quill.theme.tooltip.hide();
+        saveLink(this.quill, selection.index);
       };
     } else {
       // 当有选中文本，弹出tooltip后允许修改此文本，并能插入超链接
@@ -114,21 +110,12 @@ export function linkHandler(value) {
       if (isEmail(text) && text.indexOf('mailto:') !== 0) {
         href = `mailto:${text}`;
       } else if (isUrl(text)) {
-        if (!text.startsWith('http:') && !text.startsWith('https:')) {
-          href = `http://${text}`;
-        }
+        href = normalizeUrl(text);
       }
       document.getElementById('link-url').value = href;
       document.querySelector('a.ql-action').onclick = () => {
         this.quill.deleteText(selection.index, selection.length);
-        this.quill.insertText(
-          selection.index,
-          document.getElementById('link-words').value,
-          'link',
-          document.getElementById('link-url').value,
-          'user',
-        );
-        this.quill.theme.tooltip.hide();
+        saveLink(this.quill, selection.index);
       };
     }
   } else {
@@ -161,5 +148,6 @@ export const uploadImg = (blob, imgUploadApi, uploadSuccCB, uploadFailCB) => {
     // uploadFailCB('1234');
   } catch (e) {
     console.log('uploadImg', e);
+    uploadFailCB(e);
   }
 };