+++ rollup.config.mjs
@@ -23,8 +23,8 @@ export default {
       name: 'quillReactCommercial',
     },
     {
-      dir: path.dirname(pkg.module), // 打包输出文件保留原始模块结构
-      format: 'es',
+      dir: path.dirname(pkg.main), // 打包输出文件保留原始模块结构
+      format: 'cjs',
       name: pkg.name,
       exports: 'named', // 指定导出模式（自动、默认、命名、无）
       preserveModules: true, // 保留模块结构
@@ -42,7 +42,7 @@ export default {
     }),
     postcss(),
     // image(), // 将所有图片打包成base64
-    strip(), // Remove debugger statements and functions like assert.equal and console.log from your code
+    strip(), // Remove debugger statements and functions like assert.equal and console.log from your code 生产环境打开
     // url(), // 将所有图片打包成js文件
     svg(), // 将svg作为inline打包，dom中直接嵌入svg dom，非base64，采用image/url打成base64有部分图片展示不出来，还需要套img src标签
   ],